+++ CHANGELOG.md
@@ -21,7 +21,7 @@ Changelog
 - **Fix:** Don't quietly stub accessors for missing `Binding.Provided` bindings.
 - **Fix:** Check constructor-annotated injections when discovering scoped classes in parent graphs.
 - **Fix:** Fix `BaseDoubleCheck.isInitialized()`.
-- **Fix:** Gracefully fall back to `MessageCollector` for graph seal errors on sourceless declarations.
+- **Fix:** Gracefully fall back to `MessageCollector` for graph seal and contributed graph errors on sourceless declarations.
 
 0.4.0
 -----